chore: adds eslint cache
